requirements.txt
pyyaml
fire
gitpython
rich